requirements.txt
django
django-unfold
pyotp
pandas
matplotlib
request